Fix database error type

+++ src/main/resources/db/migration/V3__Create_and_fill_product_table.sql
@@ -9,7 +9,7 @@ CREATE TABLE sku (
   id         INTEGER PRIMARY KEY AUTOINCREMENT,
   name       TEXT     NOT NULL,
   product_id INTEGER  NOT NULL,
-  price      REAL     NOT NULL,
+  price      INTEGER  NOT NULL,
   created_at DATETIME NOT NULL,
   updated_at DATETIME NOT NULL
 )
